lw_faas/requirements.txt
Django>=4.2
celery>=5.3
django-celery-results>=2.5
pika>=1.2

pip~=25.0.1
amqp~=5.3.1
vine~=5.1.0
kombu~=5.5.4
billiard~=4.2.1
packaging~=25.0
click~=8.2.1
python-dateutil~=2.9.0.post0
click-repl~=0.3.0
click-plugins~=1.1.1.2
asgiref~=3.9.1
sqlparse~=0.5.3
wcwidth~=0.2.13
six~=1.17.0
prompt_toolkit~=3.0.51
djangorestframework~=3.16.1